pyproject.toml
[project]
name = "recruitment-assistant"
version = "0.1.0"
description = "Windows local recruitment resume collection and management assistant."
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.35.0",
    "playwright>=1.44.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg[binary]>=3.1.19",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "openpyxl>=3.1.2",
    "python-docx>=1.1.2",
    "parsel>=1.9.1",
    "beautifulsoup4>=4.12.3",
    "apscheduler>=3.10.4",
    "loguru>=0.7.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.4.4"
]

[tool.setuptools.packages.find]
include = ["recruitment_assistant*"]
exclude = ["app*", "tests*", "docs*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
